latest changes

+++ src/app/module/admin/pages/announcement/component/edit-announcement/edit-announcement.component.ts
@@ -35,6 +35,9 @@ export class EditAnnouncementComponent {
   }
 
   public deleteAnnouncement() {
+     if (!confirm('Are you sure to delete this Announcement?')) {
+      return;
+    }
     this.announcementService.deleteAnnouncement(this.announcement).subscribe({
       next: () => {
         this.closeModel();
